Migrate sam's code from tuesday alongside the rest
I've tried to put all the code near or into the relevant methods
so it can be easily integrated as needed and the rest removed.

I've dropped ReservationManager since i don't think we need a class
for it. InetSocketAddress has IP and port. and the server ID for the
mutex is buried in timestamp and clock's data scructures.

I also dropped getTheSmallestRequest since it isn't needed due to
that being part of Garg's code and the PriorityQueue