Fix naming conventions and IDEA project files
Fixing the naming conventions was never going to be easier to do
than now before we start in earnest. If we were ever to make this
public it should be easy for others to read and include in their project.

I'm hoping having this as a proper intellij IDEA project will make
it easier for us to collaborate and setup on other computers.